feat: properly render logos
