+++ packages/components/src/components/ScalarIcon/ScalarIcon.stories.ts
@@ -2,7 +2,8 @@ import type { Meta, StoryObj } from '@storybook/vue3'
 
 import IconList from './IconList.vue'
 import ScalarIcon from './ScalarIcon.vue'
-import { ICONS } from './icons/icons'
+import { ICONS } from './icons'
+import { LOGOS } from './logos'
 
 const meta = {
   component: ScalarIcon,
