+++ packages/components/src/components/ScalarIcon/ScalarIcon.vue
@@ -3,19 +3,31 @@ import type { VariantProps } from 'cva'
 import { computed } from 'vue'
 
 import { cva, cx } from '../../cva'
-import { type Icon, getIcon } from './icons/'
+import { type Icon, type Logo, getIcon, getLogo } from './utils'
+
+/**
+ * Icon wrapper for all icons and logos
+ */
 
 type IconVariants = VariantProps<typeof iconProps>
 
-/**
- * Icon wrapper for all scalar icons
- */
-const props = defineProps<{
-  icon: Icon
+type GenericProps = {
   size?: IconVariants['size']
   thickness?: string
   label?: string
-}>()
+}
+
+type IconProps = {
+  icon: Icon
+  logo?: never
+} & GenericProps
+
+type LogoProps = {
+  icon?: never
+  logo: Logo
+} & GenericProps
+
+const props = defineProps<IconProps | LogoProps>()
 
 const iconProps = cva({
   variants: {
@@ -45,10 +57,17 @@ const accessibilityAttrs = computed(() =>
         role: 'presentation',
       },
 )
+
+const svg = computed(() => {
+  if (props.icon) return getIcon(props.icon)
+  if (props.logo) return getLogo(props.logo)
+
+  return undefined
+})
 </script>
 <template>
   <component
-    :is="getIcon(icon)"
+    :is="svg"
     :class="cx('scalar-icon', iconProps({ size }))"
     v-bind="accessibilityAttrs" />
 </template>
