--- packages/components/src/components/ScalarIcon/icons/icons.ts
@@ -1,134 +1,0 @@
-export const ICONS = [
-  'Add',
-  'AddTab',
-  'Adonis',
-  'ArrowLeft',
-  'ArrowRight',
-  'Brackets',
-  'Branch',
-  'Changelog',
-  'Checkmark',
-  'ChevronDown',
-  'ChevronLeft',
-  'ChevronRight',
-  'ChevronUp',
-  'Clipboard',
-  'Close',
-  'CloseTab',
-  'CloseTabs',
-  'Cloud',
-  'CodeFolder',
-  'Cog',
-  'Collection',
-  'Cookie',
-  'DarkMode',
-  'Delete',
-  'Discord',
-  'DiscordLine',
-  'DocsPage',
-  'Dotnet',
-  'Download',
-  'Duplicate',
-  'Edit',
-  'Ellipses',
-  'Elysia',
-  'Email',
-  'Error',
-  'Example',
-  'Express',
-  'ExternalLink',
-  'Fastapi',
-  'Fastify',
-  'FilterList',
-  'Folder',
-  'GitHub',
-  'GitHubLine',
-  'Go',
-  'Google',
-  'Help',
-  'Hide',
-  'History',
-  'Hono',
-  'House',
-  'Import',
-  'Info',
-  'Key',
-  'Laravel',
-  'Leave',
-  'LightDarkModeToggle',
-  'LightMode',
-  'Link',
-  'Litestar',
-  'Lock',
-  'Logo',
-  'LogoAPI',
-  'LogoClient',
-  'LogoMarket',
-  'LogoSwagger',
-  'Magic',
-  'Menu',
-  'Nestjs',
-  'Nextjs',
-  'Nitro',
-  'Nuxt',
-  'Openapi',
-  'Page',
-  'PaperAirplane',
-  'Pencil',
-  'Platformatic',
-  'Play',
-  'Refresh',
-  'Roadmap',
-  'Rust',
-  'Scribble',
-  'Search',
-  'Server',
-  'Settings',
-  'Show',
-  'Sort',
-  'Terminal',
-  'ThumbsDown',
-  'ThumbsUp',
-  'Trash',
-  'Unlock',
-  'Unwatch',
-  'Upload',
-  'UploadSimple',
-  'User',
-  'Users',
-  'Wallet',
-  'Watch',
-  'Workspace',
-  'programming-framework-angular',
-  'programming-framework-astro',
-  'programming-framework-laravel',
-  'programming-framework-nextdotjs',
-  'programming-framework-react',
-  'programming-framework-vuedotjs',
-  'programming-language-c',
-  'programming-language-clojure',
-  'programming-language-csharp',
-  'programming-language-css3',
-  'programming-language-go',
-  'programming-language-html5',
-  'programming-language-http',
-  'programming-language-java',
-  'programming-language-javascript',
-  'programming-language-json',
-  'programming-language-kotlin',
-  'programming-language-node',
-  'programming-language-objc',
-  'programming-language-ocaml',
-  'programming-language-php',
-  'programming-language-powershell',
-  'programming-language-python',
-  'programming-language-r',
-  'programming-language-ruby',
-  'programming-language-scala',
-  'programming-language-shell',
-  'programming-language-swift',
-  'programming-language-typescript',
-  'programming-script-code',
-  'programming-tool-git',
-  'programming-tool-tailwindcss',
-] as const
