+++ packages/components/src/components/ScalarIcon/icons/index.ts
@@ -1,23 +1,123 @@
-import type { ICONS } from './icons'
-
-export type Icon = (typeof ICONS)[number]
-
-const icons = import.meta.glob<SVGElement>('./*.svg', {
-  query: 'component',
-  eager: true,
-})
-
-/**
- * Generate a vue component from the icon SVGs
- *
- * For any changes to this file, please ensure it works with SSR
- */
-export const getIcon = (name: Icon) => {
-  const filename = `./${name}.svg`
-
-  if (!icons[filename]) {
-    console.warn(`Could not find icon: ${name}`)
-    return null
-  }
-  return icons[filename]
-}
+export const ICONS = [
+  'Add',
+  'AddTab',
+  'Adonis',
+  'ArrowLeft',
+  'ArrowRight',
+  'Brackets',
+  'Branch',
+  'Changelog',
+  'Checkmark',
+  'ChevronDown',
+  'ChevronLeft',
+  'ChevronRight',
+  'ChevronUp',
+  'Clipboard',
+  'Close',
+  'CloseTab',
+  'CloseTabs',
+  'Cloud',
+  'CodeFolder',
+  'Cog',
+  'Collection',
+  'Cookie',
+  'DarkMode',
+  'Delete',
+  'Discord',
+  'DiscordLine',
+  'DocsPage',
+  'Dotnet',
+  'Download',
+  'Duplicate',
+  'Edit',
+  'Ellipses',
+  'Email',
+  'Error',
+  'Example',
+  'ExternalLink',
+  'FilterList',
+  'Folder',
+  'GitHub',
+  'GitHubLine',
+  'Google',
+  'Help',
+  'Hide',
+  'History',
+  'House',
+  'Import',
+  'Info',
+  'Key',
+  'Leave',
+  'LightDarkModeToggle',
+  'LightMode',
+  'Link',
+  'Logo',
+  'LogoAPI',
+  'LogoClient',
+  'LogoMarket',
+  'LogoSwagger',
+  'Magic',
+  'Menu',
+  'Nestjs',
+  'Nextjs',
+  'Nitro',
+  'Nuxt',
+  'Page',
+  'PaperAirplane',
+  'Pencil',
+  'Platformatic',
+  'Play',
+  'Refresh',
+  'Roadmap',
+  'Scribble',
+  'Search',
+  'Server',
+  'Settings',
+  'Show',
+  'Sort',
+  'Terminal',
+  'ThumbsDown',
+  'ThumbsUp',
+  'Trash',
+  'Unlock',
+  'Unwatch',
+  'Upload',
+  'UploadSimple',
+  'User',
+  'Users',
+  'Wallet',
+  'Watch',
+  'Workspace',
+  'programming-framework-angular',
+  'programming-framework-astro',
+  'programming-framework-laravel',
+  'programming-framework-nextdotjs',
+  'programming-framework-react',
+  'programming-framework-vuedotjs',
+  'programming-language-c',
+  'programming-language-clojure',
+  'programming-language-csharp',
+  'programming-language-css3',
+  'programming-language-go',
+  'programming-language-html5',
+  'programming-language-http',
+  'programming-language-java',
+  'programming-language-javascript',
+  'programming-language-json',
+  'programming-language-kotlin',
+  'programming-language-node',
+  'programming-language-objc',
+  'programming-language-ocaml',
+  'programming-language-php',
+  'programming-language-powershell',
+  'programming-language-python',
+  'programming-language-r',
+  'programming-language-ruby',
+  'programming-language-scala',
+  'programming-language-shell',
+  'programming-language-swift',
+  'programming-language-typescript',
+  'programming-script-code',
+  'programming-tool-git',
+  'programming-tool-tailwindcss',
+] as const
