+++ packages/components/src/components/ScalarIcon/index.ts
@@ -1,2 +1,2 @@
 export { default as ScalarIcon } from './ScalarIcon.vue'
-export { type Icon } from './icons'
+export { type Icon } from './utils'
