+++ packages/components/src/components/ScalarIcon/logos/index.ts
@@ -1,0 +1,20 @@
+export const LOGOS = [
+  'Adonisjs',
+  'Dotnet',
+  'Elysiajs',
+  'Express',
+  'Fastapi',
+  'Fastify',
+  'Go',
+  'Hono',
+  'Laravel',
+  'Litestar',
+  'Nestjs',
+  'Nextjs',
+  'Nitro',
+  'Nuxt',
+  'Platformatic',
+  'Rust',
+  'Vue',
+  'Openapi',
+] as const
